+++ Samples~/Scripts/SaintsEditor/Testing/ShowInInspectorGenImprove/AbsU1.cs
@@ -1,0 +1,6 @@
+namespace SaintsField.Samples.Scripts.SaintsEditor.Testing.ShowInInspectorGenImprove
+{
+    public class AbsU1 : AbsUnityObject
+    {
+    }
+}