+++ Samples~/Scripts/SaintsEditor/Testing/ShowInInspectorGenImprove/AbsU2.cs
@@ -1,0 +1,8 @@
+using UnityEngine;
+
+namespace SaintsField.Samples.Scripts.SaintsEditor.Testing.ShowInInspectorGenImprove
+{
+    public class AbsU2 : AbsUnityObject
+    {
+    }
+}