+++ Samples~/Scripts/SaintsEditor/Testing/ShowInInspectorGenImprove/InterObj2.cs
@@ -1,0 +1,8 @@
+using UnityEngine;
+
+namespace SaintsField.Samples.Scripts.SaintsEditor.Testing.ShowInInspectorGenImprove
+{
+    public class InterObj2 : MonoBehaviour, IObject
+    {
+    }
+}