+++ Samples~/Scripts/SaintsEditor/Testing/ShowInInspectorGenImprove/ShowInInspectorAbsUnity.cs
@@ -1,0 +1,11 @@
+using SaintsField.Playa;
+using UnityEngine;
+
+namespace SaintsField.Samples.Scripts.SaintsEditor.Testing.ShowInInspectorGenImprove
+{
+    public class ShowInInspectorAbsUnity : SaintsMonoBehaviour
+    {
+        [ShowInInspector] private Object _unityObject;
+        [ShowInInspector] private AbsUnityObject _absObject;
+    }
+}